requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.6.4
SQLAlchemy==2.0.28
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
jinja2==3.1.4
pytest==8.1.1
pytest-asyncio==0.23.6
anyio==4.3.0
